+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 - Add --script-readable to env and env docker @bexelbie
 - Fix #178: Add status command and separate status from env @bexelbie
+- Fix#173: Shows if kubernetes services is running in the box @navidshaikh
 
 ## v1.0.1 Apr 12, 2016
 - Updated SPEC (v1.0.0) for url, date and format @budhrg