+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Fix #191: 'vagrant service-manager restart' not handled correctly @budhrg
 - Fixes #187, Updated commands in the Available Commands section @preeticp
 - Fix #200: Simplify the eval hint for `vagrant service-manager env` command @budhrg
+- Add environment variables for Openshift env output @bexelbie
 
 ## v1.0.1 Apr 12, 2016
 - Updated SPEC (v1.0.0) for url, date and format @budhrg